feat: splitted format data, splitted fetch
