+++ src/network/fetchGraphData/helpers/handleTopics/index.ts
@@ -1,0 +1,36 @@
+import { NodeExtended } from '~/types'
+import { maxScale } from '../../const'
+import { TopicMap } from '../../types'
+
+export const generateTopicNodesFromMap = (topicMap: TopicMap, doNodeCallback: (node: NodeExtended) => void) => {
+  Object.entries(topicMap).forEach(([topic, content], index) => {
+    const { children, position } = content
+    const { x, y, z } = position
+    /** we dont create topic node for search term,
+     *  otherwise everything will be linked to it
+     */
+
+    const scale = children.length * 2 > maxScale ? maxScale : children.length * 2
+    const topicNodeId = `topic_node_${index}`
+
+    const topicNode: NodeExtended = {
+      x,
+      y,
+      z,
+      children,
+      colors: ['#000'],
+      id: topicNodeId,
+      label: topic,
+      name: topic,
+      type: 'topic',
+      node_type: 'topic',
+      ref_id: topicNodeId,
+      scale,
+      show_title: topic,
+      text: topic,
+      weight: 0,
+    }
+
+    doNodeCallback(topicNode)
+  })
+}
